requirements.txt
opencv-contrib-python
pillow-heif
Pillow
cryptography
pyopenssl
ultralytics
flask
